add throughput variable and method

+++ src/main/java/net/floodlightcontroller/loadbalancer/LBMember.java
@@ -41,7 +41,7 @@ public class LBMember {
     protected String memberId;
     protected String vipId;
     protected short weight;
-    protected short throughput;
+    protected double throughput;
     
     public LBMember() {
         id = String.valueOf((int) (Math.random()*10000));
@@ -56,7 +56,7 @@ public class LBMember {
         memberId = null;
         vipId = null;
         weight = -1;
-        throughput=0;
+        throughput= 0;
     }
 
     public void setMemberStatistics(ArrayList<Long> bytesIn, ArrayList<Long> bytesOut, int activeFlows) {
